% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luminophor.R
\name{luminophorOutput}
\alias{luminophorOutput}
\alias{renderLuminophor}
\title{Helper functions for using Lumino in Shiny}
\usage{
luminophorOutput(outputId, width = "100\%", height = "auto")

renderLuminophor(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{ID of htmlwidgets container}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression to create a Lumino dock htmlwidget (normally
via \code{\link{luminophor}()}).}

\item{env}{The environment in which to evaluate expr.}

\item{quoted}{Is expr a quoted expression (with \code{quote()})? This is
useful if you want to save an expression in a variable.}
}
\description{
These two functions are like most fooOutput() and renderFoo() functions in
the shiny package. The former is used to create an html container for a
Lumino dock panel, and the latter is used in the server logic to render
the Lumino dock panel.
}
\examples{
if (interactive()) {
  library(shiny)
  library(luminophor)
  shinyApp(
    ui = fluidPage(
      fluidRow(column(12, luminophorOutput('lmo', height='90vh')))
    ),
    server = function(input, output) {
      output$lmo <- renderLuminophor(
        luminophor() \%>\%
          addWidget("mywidget")
      )
    }
  )
}

}
