% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luminophor.R
\name{luminophorProxy}
\alias{luminophorProxy}
\title{Create a LuminophoR proxy object}
\usage{
luminophorProxy(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{Name of the LuminophoR htmlwidget}

\item{session}{Valid session object}
}
\value{
Proxy LuminophoR object
}
\description{
Create a LuminophoR proxy object, which can then be chained
to any of the API functions available (e.g. \code{\link{addWidget}}).
}
\examples{
if (interactive()) {
  library(shiny)
  library(luminophor)
  shinyApp(
    ui = fluidPage(
      sidebarLayout(
        sidebarPanel(
          actionButton('add', 'Add Widgets', icon = icon('plus'))
        ),
        mainPanel(
          luminophorOutput('lmo', height='90vh')
        )
      )
    ),
    server = function(input, output) {
      output$lmo <- renderLuminophor(
        luminophor()
      )

      observeEvent(input$add, {
        luminophorProxy('lmo') \%>\%
          addWidget("mywidget")
      })
    }
  )
}

}
\seealso{
\code{\link{luminophor}}
}
