% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosphorr.R
\name{maximizeWidget}
\alias{maximizeWidget}
\title{Maximize a widget in the dock}
\usage{
maximizeWidget(proxy, widgetId)
}
\arguments{
\item{proxy}{Proxy phosphorr object}

\item{widgetId}{ID for phosphorr widget}
}
\value{
phosphorrProxy
}
\description{
Maximize a widget in the dock
}
