% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosphorr.R
\name{minimizeWidget}
\alias{minimizeWidget}
\title{Minimize a widget in the dock}
\usage{
minimizeWidget(proxy, widgetId)
}
\arguments{
\item{proxy}{Proxy phosphorr object}

\item{widgetId}{ID for phosphorr widget}
}
\value{
phosphorrProxy
}
\description{
Minimize a widget in the dock
}
