% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosphorr.R
\docType{package}
\name{PhosphorR-package}
\alias{PhosphorR-package}
\title{PhosphorR: HTMLWidget implementation of PhosphorJS layout manager}
\description{
PhosphorR: HTMLWidget implementation of PhosphorJS layout manager
}
\examples{
if (interactive()) {
  library(shiny)
  library(phosphorr)
  shinyApp(
    ui = fluidPage(
      titlePanel("Old Faithful Geyser Data"),
      fluidRow(column(12, phosphorrOutput('pjs', height='90vh')))
    ),
    server = function(input, output) {
      output$pjs <- renderPhosphorr(
        phosphorr() \%>\%
          addWidget("widget-slider",
                    title = "Slider",
                    ui = sliderInput("bins", "Number of bins:",
                                     min = 1, max = 50, value = 30)) \%>\%
          addWidget("widget-plot",
                    title = "Plot",
                    insertmode = "split-right",
                    refwidgetID = "widget-slider",
                    relsize = 0.75,
                    ui = plotOutput("distPlot"))
      )

      output$distPlot <- renderPlot({
        x <- faithful[, 2]
        bins <- seq(min(x), max(x), length.out = input$bins + 1)
        hist(x, breaks = bins, col = 'darkgray', border = 'white')
      })
    }
  )
}
}
\seealso{
\code{\link{phosphorr}}, \code{\link{phosphorrOutput}}
}
