% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosphorr.R
\name{phosphorrOutput}
\alias{phosphorrOutput}
\alias{renderPhosphorr}
\title{Helper functions for using PhosphorJS in Shiny}
\usage{
phosphorrOutput(outputId, width = "100\%", height = "auto")

renderPhosphorr(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{ID of htmlwidgets container}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression to create a PhosphorJS dock htmlwidget (normally
via \code{\link{phosphorr}()}).}

\item{env}{The environment in which to evaluate expr.}

\item{quoted}{Is expr a quoted expression (with \code{quote()})? This is
useful if you want to save an expression in a variable.}
}
\description{
These two functions are like most fooOutput() and renderFoo() functions in
the shiny package. The former is used to create an html container for a
PhosphorJS dock panel, and the latter is used in the server logic to render
the PhosphorJS dock panel.
}
\examples{
if (interactive()) {
  library(shiny)
  library(phosphorr)
  shinyApp(
    ui = fluidPage(
      fluidRow(column(12, phosphorrOutput('pjs', height='90vh')))
    ),
    server = function(input, output) {
      output$pjs <- renderPhosphorr(
        phosphorr() \%>\%
          addWidget("mywidget")
      )
    }
  )
}

}
