% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosphorr.R
\name{phosphorrProxy}
\alias{phosphorrProxy}
\title{Create a PhosphorR proxy object}
\usage{
phosphorrProxy(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{Name of the PhosphorR htmlwidget}

\item{session}{Valid session object}
}
\value{
Proxy PhosphorR object
}
\description{
Create a PhosphorR proxy object, which can then be chained
to any of the API functions available (e.g. \code{\link{addWidget}}).
}
\examples{
if (interactive()) {
  library(shiny)
  library(phosphorr)
  shinyApp(
    ui = fluidPage(
      sidebarLayout(
        sidebarPanel(
          actionButton('add', 'Add Widgets', icon = icon('plus'))
        ),
        mainPanel(
          phosphorrOutput('pjs', height='90vh')
        )
      )
    ),
    server = function(input, output) {
      output$pjs <- renderPhosphorr(
        phosphorr()
      )

      observeEvent(input$add, {
        phosphorrProxy('pjs') \%>\%
          addWidget("mywidget")
      })
    }
  )
}

}
\seealso{
\code{\link{phosphorr}}
}
