% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosphorr.R
\name{\%>\%}
\alias{\%>\%}
\title{PhosphorR pipe operator}
\arguments{
\item{lhs, rhs}{A \link[=phosphorrProxy]{proxy PhosphorR object}}
}
\description{
Like dplyr, PhosphorR also uses the pipe function, \code{\%>\%} to turn
function composition into a series of imperative statements.
}
\examples{
# Instead of
addWidget(phosphorr(), "mywidget")
# you can write
phosphorr() \%>\% addWidget("mywidget")
}
