% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosphorr.R
\name{removeWidgets}
\alias{removeWidgets}
\title{Remove widgets in the dock}
\usage{
removeWidgets(proxy, widgetIDs = NULL, .all = FALSE)
}
\arguments{
\item{proxy}{Proxy phosphorr object}

\item{widgetIDs}{IDs for phosphorr widget}

\item{.all}{Remove all widgets?  This will override any ids specified in \code{widgetIDs}}
}
\value{
phosphorrProxy
}
\description{
Remove widgets in the dock
}
