% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosphorr.R
\name{widgetBody}
\alias{widgetBody}
\title{Create a body for a phosphorr widget}
\usage{
widgetBody(..., .list = NULL)
}
\arguments{
\item{...}{Items to put in the header}

\item{.list}{An optional list containing items to put in the header. Same as the ... arguments, but in list format. This can be useful when working with programmatically generated items.}
}
\value{
shiny.tag
}
\description{
Create a body for a phosphorr widget
}
