% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosphorr.R
\name{widgetHeader}
\alias{widgetHeader}
\title{Create a header for a phosphorr widget}
\usage{
widgetHeader(..., disable = FALSE, .list = NULL)
}
\arguments{
\item{...}{Items to put in the header}

\item{disable}{If TRUE, don't display the header bar.}

\item{.list}{An optional list containing items to put in the header. Same as the ... arguments, but in list format. This can be useful when working with programmatically generated items.}
}
\value{
shiny.tag
}
\description{
Create a header for a phosphorr widget
}
