% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosphorr.R
\name{addWidget}
\alias{addWidget}
\title{Adds a new widget to the dock}
\usage{
addWidget(phosphorrProxy, id, title = "Widget", closable = TRUE,
  content = "", ui = HTML("I am a widget!"))
}
\arguments{
\item{phosphorrProxy}{Proxy phosphorr object}

\item{id}{ID for phosphorr widget}

\item{content}{Code for phosphorr widget UI}

\item{ui}{Shiny UI content.  If just text, need to use HTML(...)}
}
\value{
phosphorrProxy
}
\description{
Adds a new widget to the dock
}
