% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luminophor.R
\name{getLayout}
\alias{getLayout}
\title{Get layout of the dock}
\usage{
getLayout(proxy)
}
\arguments{
\item{proxy}{\link[=luminophorProxy]{Proxy LuminophoR object}}
}
\description{
Get layout of the dock
}
