% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phosphorr.R
\name{phosphorrProxy}
\alias{phosphorrProxy}
\title{Create a phosphorr proxy object}
\usage{
phosphorrProxy(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{Name of the phosphorr panel}

\item{session}{Valid session object}
}
\value{
panel proxy object
}
\description{
Create a phosphorr proxy object
}
