% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luminophor.R
\name{\%>\%}
\alias{\%>\%}
\title{LuminophoR pipe operator}
\arguments{
\item{lhs, rhs}{A \link[=luminophorProxy]{proxy LuminophoR object}}
}
\description{
Like dplyr, LuminophoR also uses the pipe function, \code{\%>\%} to turn
function composition into a series of imperative statements.
}
\examples{
\dontrun{
# Instead of
addWidget(luminophor(), "mywidget")
# you can write
luminophor() \%>\% addWidget("mywidget")
}
}
